+++ engine/src/main/java/net/jqwik/engine/properties/configurators/StringLengthConfigurator.java
@@ -4,15 +4,37 @@ import net.jqwik.api.*;
 import net.jqwik.api.arbitraries.*;
 import net.jqwik.api.configurators.*;
 import net.jqwik.api.constraints.*;
+import net.jqwik.api.providers.*;
 
 public class StringLengthConfigurator extends ArbitraryConfiguratorBase {
 
-	public StringArbitrary configure(StringArbitrary arbitrary, StringLength stringLength) {
+	@Override
+	protected boolean acceptTargetType(TypeUsage targetType) {
+		return targetType.isOfType(String.class);
+	}
+
+	public Arbitrary<String> configure(Arbitrary<String> arbitrary, StringLength stringLength) {
 		checkSize(stringLength);
-		if (stringLength.value() != 0) {
-			return arbitrary.ofLength(stringLength.value());
+		if (arbitrary instanceof StringArbitrary) {
+			return configureStringArbitrary((StringArbitrary) arbitrary, stringLength);
 		} else {
-			StringArbitrary newArbitrary = arbitrary;
+			return configureOtherArbitrary(arbitrary, stringLength);
+		}
+	}
+
+	private Arbitrary<String> configureOtherArbitrary(Arbitrary<String> arbitrary, StringLength stringLength) {
+		if (stringLength.value() != 0) {
+			return arbitrary.filter(s -> s.length() == stringLength.value());
+		} else {
+			return arbitrary.filter(s -> s.length() >= stringLength.min() && s.length() <= stringLength.max());
+		}
+	}
+
+	private Arbitrary<String> configureStringArbitrary(StringArbitrary stringArbitrary, StringLength stringLength) {
+		if (stringLength.value() != 0) {
+			return stringArbitrary.ofLength(stringLength.value());
+		} else {
+			StringArbitrary newArbitrary = stringArbitrary;
 			if (stringLength.min() != 0) {
 				newArbitrary = newArbitrary.ofMinLength(stringLength.min());
 			}